feat(ac4): add object playback and IMS investigation tools
Integrate pinned MacinDecode Full A-JOC decoding, MP4 transport and bounded streaming. Add a diagnostic stereo exporter and reproducible IMS metadata probes; IMS spatial rendering remains unimplemented.

+++ packages/core/src/eac3_pipeline.rs
@@ -493,7 +493,7 @@ fn extract_events(
                     sample_pos: base_sample_pos + sample_offset,
                     has_pos,
                     pos,
-                    gain_db: object_gain_db(block.gain, block.inactive),
+                    gain_db: f64::from(object_gain_db(block.gain, block.inactive)),
                     size,
                     anchor,
                     distance_m,
